Checkout System added

+++ shop/views/api.py
@@ -1,53 +1,67 @@
 from django.http import JsonResponse, HttpResponse
 from django.contrib.auth.models import User
 from django.db.models import Sum
-from ..models import Cart, CartItem
+from ..models import Cart, CartItem, Customer
 from pprint import pprint
 
+
 def getCartInfo(request):
-    user_id = request.user.id
-    
-    if user_id is not None:
-        cart = Cart.objects.filter(user_id=user_id).first()
+    if request.user.is_authenticated and hasattr(request.user, 'customer'):
+        customer_id = request.user.customer.id
+        cart = Cart.objects.filter(customer_id=customer_id).first()
         if cart is not None and hasattr(cart, 'items'):
-            items = list(cart.items.all().values('id', 'cart_id', 'product__name', 'quantity', 'product__image', 'product__price'))
+            items = list(cart.items.all().values(
+                'id', 'cart_id', 'product__name', 'quantity', 'product__image', 'product__price'))
             count = list(cart.items.all().aggregate(Sum('quantity')).values())
             res = {
                 'count': count[0],
                 'items': items
             }
             return JsonResponse(res, safe=False)
-    
+
     return JsonResponse([], safe=False)
 
+
 def addToCart(request):
-    user_id = request.user.id
-    cart = Cart.objects.filter(user_id=user_id).first()
-    product_id = request.POST.get('product_id')
-    quantity = request.POST.get('quantity')
+    user = request.user
+    if user.is_authenticated:
+        if not hasattr(user, 'customer'):
+            Customer.objects.create(user=user)
+        customer_id = user.customer.id
+        cart = Cart.objects.filter(customer_id=customer_id).first()
+        product_id = request.POST.get('product_id')
+        quantity = request.POST.get('quantity')
 
-    if cart is None:
-        cart = Cart.objects.create(user_id=user_id)
-        
-    CartItem.objects.create(cart_id=cart.id, product_id=product_id, quantity=quantity)
+        if int(quantity) > 0:
+            if cart is None:
+                cart = Cart.objects.create(customer_id=customer_id)
+            CartItem.objects.create(
+                cart_id=cart.id, product_id=product_id, quantity=quantity)
+            return HttpResponse("success")
 
-    return HttpResponse("success")
+    return HttpResponse("fail")
 
 
 def updateCart(request):
-    cartitem_id = request.POST.get('cartitem_id')
-    quantity = request.POST.get('quantity')
+    if request.user.is_authenticated:
+        cartitem_id = request.POST.get('cartitem_id')
+        quantity = request.POST.get('quantity')
+        if int(quantity) > 0:
+            cartitem = CartItem.objects.get(pk=cartitem_id)
+            cartitem.quantity = quantity
+            cartitem.save(update_fields=['quantity'])
+            return HttpResponse("success")
 
-    cartitem = CartItem.objects.get(pk=cartitem_id)
-    cartitem.quantity = quantity
-    cartitem.save(update_fields = ['quantity'])
+    return HttpResponse("failed")
 
-    return HttpResponse("success")
 
 def deleteCart(request):
-    cartitem_id = request.POST.get('cartitem_id')
+    if request.user.is_authenticated:
+        cartitem_id = request.POST.get('cartitem_id')
 
-    cartitem = CartItem.objects.get(pk=cartitem_id)
-    cartitem.delete() 
+        cartitem = CartItem.objects.get(pk=cartitem_id)
+        cartitem.delete()
 
-    return HttpResponse("success")
+        return HttpResponse("success")
+
+    return HttpResponse("fail")
